tests: Moved hamsterlib related test into own dir.
