+++ tests/backends/sqlalchemy/common.py
@@ -1,0 +1,2 @@
+from sqlalchemy import orm
+Session = orm.scoped_session(orm.sessionmaker())
