+++ tests/hamsterlib/test_helpers.py
