+++ tests/hamsterlib/test_lib.py
