+++ tests/hamsterlib/test_objects.py
