+++ tests/hamsterlib/test_reports.py
