+++ tests/hamsterlib/test_storage.py
